Fix int to long for flags_stollen

+++ src/ctf01d/objects/ctf01d_scoreboard.cpp
@@ -300,7 +300,7 @@ std::optional<int> Ctf01dScoreboard::incrementAttackScore(const ctf01d::flag &fl
   // TODO calculate
   // int nFlagPoints = m_mapServiceCostsAndStatistics[sServiceId]->getCostStolenFlag()*10; // one number after dot
   int flag_points = m_flag_cost_in_points->value(); // TODO basic
-  int nDateAction = WsjcppCore::getCurrentTimeInMilliseconds();
+  long nDateAction = WsjcppCore::getCurrentTimeInMilliseconds();
   // victim place in scoreboard
   std::map<std::string, Ctf01dTeamStatusRow *>::iterator it_victim;
   it_victim = m_mapTeamsStatuses.find(flag.getTeamId());
